solved all linter issues
