--- DOMelements.js
@@ -1,6 +1,0 @@
-export const endModal = document.querySelector('#end-modal');
-export const cells = [...document.querySelectorAll('.cell')];
-export const grid = document.querySelector('.board');
-export const startModal = document.querySelector('#start-modal');
-export const playerOneInput = startModal.querySelector('#p1');
-export const playerTwoInput = startModal.querySelector('#p2');
