--- factories.js
@@ -1,6 +1,0 @@
-const playerFactory = (name, mark) => ({
-  name,
-  mark,
-});
-
-export default playerFactory;
