+++ js/factories.js
@@ -1,0 +1,6 @@
+const playerFactory = (name, mark) => ({
+  name,
+  mark,
+});
+
+export default playerFactory;
